+++ OnlineBookReader.cpp
@@ -1,755 +1,7 @@
 #include <bits/stdc++.h>
+#include "admins.h"
 #define sp " "
 using namespace std;
-
-int safe_input_int(int min, int max)
-{
-    int choice;
-    while (true)
-    {
-        cin >> choice;
-        if (cin.fail())
-        {
-            cout << "Invalid input! Expected an integer.\n";
-            cout << "Please try again: ";
-            cin.clear();
-            cin.ignore(numeric_limits<streamsize>::max(), '\n');
-            continue;
-        }
-        if (choice < min || choice > max)
-        {
-            cout << "Out of range, please try again.\n";
-            cout << "Please try again: ";
-            cin.clear();
-            cin.ignore(numeric_limits<streamsize>::max(), '\n');
-            continue;
-        }
-        
-        return choice;
-    }
-}
-
-class book
-{
-private:
-    string ISBN;
-    string title;
-    string author_name;
-    int cnt_of_pages;
-    vector<string> pages;
-public:
-    book()
-    {
-        ISBN = "";
-        title = "";
-        author_name = "";
-        cnt_of_pages = 0;    
-    }
-
-    void set_ISBN(string s)
-    {
-        ISBN = s;
-    }
-    void set_title(string s)
-    {
-        title = s;
-    }
-    void set_author_name(string s)
-    {
-        author_name = s;
-    }
-    void set_cnt_of_pages(int n)
-    {
-        cnt_of_pages = n;
-    }
-    void set_pages(vector<string> v)
-    {
-        pages = v;
-    }
-    
-    string get_ISBN()
-    {
-        return ISBN;
-    }
-    string get_title()
-    {
-        return title;
-    }
-    string get_author_name()
-    {
-        return author_name;
-    }
-    int get_cnt_of_pages()
-    {
-        return cnt_of_pages;
-    }
-    vector<string> get_pages()
-    {
-        return pages;
-    }
-
-    void add_book()
-    {
-        // cin.ignore();
-        cout << "Enter ISBN: ";
-        string str;
-        getline(cin,str);
-        set_ISBN(str);
-        cout << "Enter Title: ";
-        getline(cin,str);
-        set_title(str);
-        cout << "Enter Author Name: ";
-        getline(cin,str);
-        set_author_name(str);
-        cout << "Enter How many pages: ";
-        int n;
-        cin >> n;
-        cin.ignore();
-        set_cnt_of_pages(n);
-        vector<string> entered_pages;
-        for(int i = 0; i < n; i++)
-        {
-            cout << "Enter page # " << i+1 << ": ";  
-            getline(cin,str);
-            pages.push_back(str);
-        }
-        cout << "Book added successfully.\n\n";
-    }
-
-};
-vector<book> all_books;
-void books_in_the_system()
-{
-    book b;
-    ifstream fin("books_in_the_system.txt");
-    string line;
-    while(getline(fin,line))
-    {
-        book b;
-        stringstream ss(line);
-
-        string title, isbn, author;
-        string pages_count_str;
-        int pages_count;
-
-        getline(ss, title, ',');
-        getline(ss, isbn, ',');
-        getline(ss, author, ',');
-        getline(ss, pages_count_str, ','); 
-        
-        pages_count = stoi(pages_count_str);
-
-        b.set_title(title);
-        b.set_ISBN(isbn);
-        b.set_author_name(author);
-        b.set_cnt_of_pages(pages_count);
-        
-        vector<string> book_pages;
-        string page;
-        for (int i = 0; i < pages_count; i++)
-        {
-            if (i < pages_count - 1)
-            {
-                getline(ss, page, ',');
-            }
-            else
-            {
-                getline(ss, page);
-            }
-            book_pages.push_back(page);
-        }
-
-        b.set_pages(book_pages);        
-        all_books.push_back(b);
-
-    }
-}
-
-
-set<string> all_usernames;
-class user
-{
-private:
-    string username;
-    string password;
-    string email;
-    string name;
-    vector<tuple<string,int,string>> sessions;
-    unordered_map<string, int> book_current_page;
-    bool taken;
-
-public:
-    user()
-    {
-        username = "";
-        password = "";
-        email = "";
-        name = "";
-        taken = false;
-    }
-
-    void set_username(string s)
-    {
-        username = s;
-    }
-    void set_password(string s)
-    {
-        password = s;
-    }
-    void set_email(string s)
-    {
-        email = s;
-    }
-    void set_name(string s)
-    {
-        name = s;
-    }
-
-    string get_username()
-    {
-        return username;
-    }
-    string get_password()
-    {
-        return password;
-    }
-    string get_email()
-    {
-        return email;
-    }
-    string get_name()
-    {
-        return name;
-    }
-    string get_current_time_date() 
-    {
-	    // src: https://stackoverflow.com/questions/17223096/outputting-date-and-time-in-c-using-stdchrono
-        auto now = std::chrono::system_clock::now();
-	    auto in_time_t = std::chrono::system_clock::to_time_t(now);
-	    std::stringstream ss;
-	    ss << std::put_time(std::localtime(&in_time_t), "%Y-%m-%d %X");
-        return ss.str();
-    }
-    
-    void read()
-    {
-        cin.ignore();
-        cout << "Enter username: ";
-        string username;
-        getline(cin,username);
-
-        cout << "Enter password: ";
-        string password;
-        getline(cin,password);
-
-        cout << "Enter name: ";
-        string name;
-        getline(cin,name);
-
-        cout << "Enter email: ";
-        string email;
-        getline(cin,email);
-        if(all_usernames.count(username))
-        {
-            taken = true;
-            return void(cout << "This username is already taken, please try another one.\n\n");
-        }
-
-        set_username(username);
-        set_password(password);
-        set_name(name);
-        set_email(email);
-    }
-
-    void view_profile()
-    {
-        cout << endl;
-        cout << "Name: " << get_name() << endl;
-        cout << "Email: " << get_email() << endl;
-        cout << "Username: " << get_username() << endl;
-    }
-    
-    void pages_menu(book bk)
-    {
-        string book_id = bk.get_title();
-        int& curr = book_current_page[book_id];
-        if(curr == 0)
-            curr = 1;
-        cout << "Menu:\n";
-        cout << "1: Next page\n";
-        cout << "2: Previous page\n";
-        cout << "3: Stop reading\n";
-        cout << "Enter number in rage(1-3): ";
-        int choice = safe_input_int(1,3);  
-        
-        if(choice == 1)
-        {
-            if(curr == bk.get_cnt_of_pages())
-            {
-                cout << "You are already in the last page.\n";
-                print_current_page(bk);
-            }
-            else
-            {
-                curr++;
-                print_current_page(bk);
-            }
-        }
-        else if(choice == 2)
-        {
-            if(curr == 1)
-            {
-                cout << "You are already in the first page.\n\n";
-                print_current_page(bk);
-            }
-            else
-            {
-                curr--;
-                print_current_page(bk);
-            }
-        }
-        else
-        {
-            cout << "Stopping reading " << bk.get_title() << ".\n";
-            sessions.push_back({bk.get_title(), bk.get_cnt_of_pages(), get_current_time_date()});
-            return;
-        }
-    }
-
-    void print_current_page(book bk)
-    {
-        string book_id = bk.get_title();
-        int curr = book_current_page[book_id];
-        if(curr == 0)
-            curr = 1;
-
-        cout << "\nCurrent Page: " << curr << "\\" << bk.get_cnt_of_pages() << endl;
-        cout << bk.get_pages()[curr-1] << "\n\n";
-        pages_menu(bk);
-    }
-    
-    book b;
-    void books_form_the_system()
-    {
-        cout << "\nOur current book collection:\n";
-        for (int i = 1; i <= all_books.size(); i++)
-        {
-            cout << "  " << i << ") " << all_books[i - 1].get_title() << endl;
-        }
-        cout << "\nWhich Book you want to read?\n";
-        cout << "Enter number in range (1-" << all_books.size() << "): ";
-        int choice = safe_input_int(1, all_books.size());
-        b = all_books[choice - 1];
-        print_current_page(b);
-    }
-
-    void books_form_my_history()
-    {
-        if (sessions.size() == 0)
-        {
-            return void(cout << "You have no reading history yet, please read a book from the system first.\n\n");
-        }
-        for (int i = 1; i <= sessions.size(); i++)
-        {
-            int curr = book_current_page[get<0>(sessions[i - 1])];
-            if (curr == 0)
-                curr = 1;
-            cout << i << ") " << get<0>(sessions[i - 1])
-                 << " (" << curr << "\\" << get<1>(sessions[i - 1])
-                 << ") | Last read at: " << get<2>(sessions[i - 1]) << endl;
-        }
-
-        cout << "\nWhich session you want to open?\n";
-        cout << "Enter number in range (1-" << sessions.size() << "): ";
-
-        int choice = safe_input_int(1, sessions.size());
-        bool found = false;
-        for (int i = 0; i < all_books.size(); i++)
-        {
-            if (get<0>(sessions[choice - 1]) == all_books[i].get_title())
-            {
-                b = all_books[i];
-                found = true;
-                print_current_page(b);
-                break;
-            }
-        }
-        if (!found)
-        {
-            cout << "Book from this session not found in library.\n";
-        }
-    }
-
-    void menu()
-    {
-        if(taken)
-            return;
-        cout << "\nHello " << get_name() << " | User view";
-        while(true)
-        {
-            cout << "\nMenu:\n";
-            cout << "1: View profile\n";
-            cout << "2: List & Select from My Reading History\n";
-            cout << "3: List & Select from Available Books\n";
-            cout << "4: Logout\n";
-            cout << "\nEnter a number in range (1-4): ";
-            int choice = safe_input_int(1,4);
-            // cin.ignore();
-            if(choice == 1)
-            {
-                view_profile();
-            }
-            else if(choice == 2)
-            {
-                books_form_my_history();
-            }
-            else if(choice == 3)
-            {
-                books_form_the_system();
-            }
-            else if(choice == 4)
-            {
-                cout << "Exiting the online book reader system.\n" << endl;
-                break;
-            }
-        }
-    }
-};
-vector<user> all_users;
-
-
-class admin
-{
-private:
-    string username;
-    string password;
-    string email;
-    string name;
-
-public:
-    admin()
-    {
-        username = "";
-        password = "";
-        email = "";
-        name = "";
-    }
-    void set_username(string s)
-    {
-        username = s;
-    }
-    void set_password(string s)
-    {
-        password = s;
-    }
-    void set_email(string s)
-    {
-        email = s;
-    }
-    void set_name(string s)
-    {
-        name = s;
-    }
-
-    string get_username()
-    {
-        return username;
-    }
-    string get_password()
-    {
-        return password;
-    }
-    string get_email()
-    {
-        return email;
-    }
-    string get_name()
-    {
-        return name;
-    }
-    
-    void read()
-    {
-        cout << "Enter username: ";
-        string username;
-        getline(cin,username);
-
-        cout << "Enter password: ";
-        string password;
-        getline(cin,password);
-
-        cout << "Enter name: ";
-        string name;
-        getline(cin,name);
-
-        cout << "Enter email: ";
-        string email;
-        getline(cin,email);
-        if(all_usernames.count(username))
-        {
-            return void(cout << "This username is already taken, please try another one.\n\n");
-        }
-
-        set_username(username);
-        set_password(password);
-        set_name(name);
-        set_email(email);
-    }
-
-    void view_profile()
-    {
-        cout << endl;
-        cout << "Name: " << get_name() << endl;
-        cout << "Email: " << get_email() << endl;
-        cout << "Username: " << get_username() << endl;
-        cout << "Password: " << get_password() << "\n\n";
-    }
-
-    void add_book()
-    {
-        book b;
-        b.add_book();
-        all_books.push_back(b);
-        ofstream fout("books_in_the_system.txt", ios::app);
-        fout << b.get_title() << "," << b.get_ISBN() << "," << b.get_author_name() << "," << b.get_cnt_of_pages();
-        for(int i = 0; i < b.get_cnt_of_pages(); i++)
-        {
-            fout << "," << b.get_pages()[i];
-        }
-        fout << "\n";
-        fout.close();
-    }
-
-    void remove_user()
-    {
-        cout << "Enter Username: ";
-        // cin.ignore();
-        string str;
-        getline(cin, str);
-        for (auto it = all_users.begin(); it != all_users.end(); ++it)
-        {
-            if (it->get_username() == str)
-            {
-                all_users.erase(it);
-                all_usernames.erase(str);
-                cout << "User '" << str << "' removed successfully.\n\n";
-                return;
-            }
-        }
-        cout << "User '" << str << "' not found.\n\n";
-    }
-
-    bool remove_book_by_title(const string& title_to_remove) 
-    {
-        ifstream fin("books_in_the_system.txt");
-        if (!fin)
-        {
-            cerr << "Error opening file for reading!" << endl;
-            return false;
-        }
-
-        vector<string> lines;
-        string line;
-
-        bool found = false;
-        while (getline(fin, line))
-        {
-            string current_title = line.substr(0, line.find(','));
-
-            if (current_title != title_to_remove)
-            {
-                lines.push_back(line);
-            }
-            else
-            {
-                found = true; 
-            }
-        }
-        fin.close();
-
-        if (!found)
-        {
-            cout << "Book with title \"" << title_to_remove << "\" not found.\n\n";
-            return false;
-        }
-
-        ofstream fout("books_in_the_system.txt");
-        if (!fout)
-        {
-            cerr << "Error opening file for writing!\n\n";
-            return false;
-        }
-
-        for (const string &remaining_line : lines)
-        {
-            fout << remaining_line << endl;
-        }
-        fout.close();
-        for (auto it = all_books.begin(); it != all_books.end(); it++)
-        {
-            if (it->get_title() == title_to_remove)
-            {
-                all_books.erase(it);
-                break; 
-            }
-        }
-        cout << "Book with title \"" << title_to_remove << "\" successfully removed.\n\n";
-        return true;
-    }
-    
-    void show_all_users()
-    {
-        cout << string(80, '-') << endl;
-        cout << left << setw(18) << "Name"
-             << setw(18) << "Username"
-             << setw(30) << "Email"
-             << setw(20) << "Password" << endl;
-        cout << string(80, '-') << endl;
-
-        // Print table rows
-        for (auto &u : all_users)
-        {
-            cout << left << setw(18) << u.get_name()
-                 << setw(18) << u.get_username()
-                 << setw(30) << u.get_email()
-                 << setw(20) << u.get_password() << endl;
-        }
-        cout << string(80, '-') << endl;
-        cout << endl;
-    }
-    
-    void show_all_books()
-    {
-        cout << string(80, '-') << endl;
-        cout << left << setw(30) << "Title"
-             << setw(20) << "Author"
-             << setw(15) << "ISBN"
-             << setw(10) << "Pages" << endl;
-        cout << string(80, '-') << endl;
-
-        for (auto &b : all_books)
-        {
-            cout << left << setw(30) << b.get_title()
-                 << setw(20) << b.get_author_name()
-                 << setw(15) << b.get_ISBN()
-                 << setw(10) << b.get_cnt_of_pages() << endl;
-        }
-        cout << string(80, '-') << endl;
-        cout << endl;
-    }
-
-};
-vector<admin> all_admins;
-void admins_in_the_system()
-{
-    string line;
-    ifstream fin("admins_in_the_system.txt");
-    while(getline(fin,line))
-    {
-        admin ad;
-        stringstream ss(line);
-
-        string name, password, email, username;
-
-        getline(ss, name, ',');
-        getline(ss, password, ',');
-        getline(ss, email, ',');
-        getline(ss, username, ',');
-
-        ad.set_name(name);
-        ad.set_password(password);
-        ad.set_email(email);
-        ad.set_username(username);
-
-        all_admins.push_back(ad);
-        all_usernames.insert(username);
-    }
-    fin.close();
-}
-
-void admin_menu(admin ad)
-{
-    cout << "\nHello " << ad.get_name() << " | Admin view\n";
-    while(true)
-    {
-        cout << "Menu:\n";
-        cout << "1: View profile\n";
-        cout << "2: Add a book\n";
-        cout << "3: Add an admin\n";
-        cout << "4: Remove a book\n";
-        cout << "5: Remove a user\n";
-        cout << "6: Show all users\n";
-        cout << "7: Show all books\n";
-        cout << "8: Show all admins\n";
-        cout << "9: Logout\n";
-        cout << "Enter a number in range (1-9): ";
-        int choice = safe_input_int(1,9);
-        cin.ignore();
-        if(choice == 1)
-        {
-            ad.view_profile();
-        }
-        else if(choice == 2)
-        {
-            ad.add_book(); 
-        }
-        else if(choice == 3)
-        {
-            admin new_admin;
-            new_admin.read();
-            all_admins.push_back(new_admin);
-            all_usernames.insert(new_admin.get_username());
-            ofstream fout("admins_in_the_system.txt", ios::app);
-            fout << new_admin.get_name() << "," << new_admin.get_password() << "," << new_admin.get_email() << "," << new_admin.get_username() << "\n";
-            fout.close();
-            cout << "Admin added successfully.\n\n";
-        }
-        else if(choice == 4)
-        {
-            cout << "Enter the title of the book to remove: ";
-            string title_to_remove;
-            getline(cin, title_to_remove);
-            ad.remove_book_by_title(title_to_remove);
-        }
-        else if(choice == 5)
-        {
-            ad.remove_user();
-        }
-        else if(choice == 6)
-        {
-            ad.show_all_users();
-        }
-        else if(choice == 7)
-        {
-            ad.show_all_books();
-        }
-        else if(choice == 8)
-        {
-            cout << string(80, '-') << endl;
-            cout << left << setw(18) << "Name"
-                 << setw(18) << "Username"
-                 << setw(30) << "Email"
-                 << setw(20) << "Password" << endl;
-            cout << string(80, '-') << endl;
-
-            for (auto &a : all_admins)
-            {
-                cout << left << setw(18) << a.get_name()
-                     << setw(18) << a.get_username()
-                     << setw(30) << a.get_email()
-                     << setw(20) << a.get_password() << endl;
-            }
-            cout << string(80, '-') << endl;
-            cout << endl;
-        }
-        
-        else if(choice == 9)
-        {
-            cout << "Logged out.\n\n";
-            break;
-        }
-    }
-}
 
 void Sign_up()
 {
